Split config and modules for driver install without litex
