+++ src/litexcnc/firmware/modules/pwm.py
@@ -2,14 +2,6 @@
 
 # Default imports
 import math
-import os
-try:
-    from typing import ClassVar, List, Literal
-except ImportError:
-    # Imports for Python <3.8
-    from typing import ClassVar, List
-    from typing_extensions import Literal
-from typing import ClassVar, List, Literal
 
 # Imports for the configuration
 from pydantic import Field
@@ -18,12 +10,9 @@ from pydantic import Field
 from migen import *
 from migen.genlib.cdc import MultiReg
 from litex.build.generic_platform import Pins, IOStandard
-from litex.soc.interconnect.csr import AutoCSR, CSRStatus, CSRStorage
+from litex.soc.interconnect.csr import AutoCSR, CSRStorage
 from litex.soc.integration.doc import AutoDoc, ModuleDoc
 from litex.soc.integration.soc import SoC
-
-# Import of the basemodel, required to register this module
-from . import ModuleBaseModel, ModuleInstanceBaseModel
 
 
 class PwmPdmModule(Module, AutoDoc, AutoCSR):
@@ -184,79 +173,3 @@ class PwmPdmModule(Module, AutoDoc, AutoCSR):
                 _pwm.period.eq(getattr(soc.MMIO_inst, f'pwm_{index}_period').storage),
                 _pwm.width.eq(getattr(soc.MMIO_inst, f'pwm_{index}_width').storage)
             ]
-
-
-class PWM_Instance(ModuleInstanceBaseModel):
-    """
-    Model describing a pin of the GPIO.
-    """
-    pin: str = Field(
-        description="The pin on the FPGA-card."
-    )
-    name: str = Field(
-        None,
-        description="The name of the pin as used in LinuxCNC HAL-file (optional). "
-        "When not specified, the standard pins like litexcnc.pwm.x will be created."
-    )
-    io_standard: str = Field(
-        "LVCMOS33",
-        description="The IO Standard (voltage) to use for the pin."
-    )
-    pins: ClassVar[List[str]] = [
-        'curr_dc',
-        'curr_period',
-        'curr_pwm_freq',
-        'curr_width',
-        'dither_pwm',
-        'enable',
-        'max_dc',
-        'min_dc',
-        'offset',
-        'pwm_freq',
-        'scale',
-        'value'
-    ]
-
-
-class PWM_ModuleConfig(ModuleBaseModel):
-    """
-    Module describing the PWM module
-    """
-    module_type: Literal['pwm'] = 'pwm'
-    module_id: ClassVar[int] = 0x70776d5f  # The string `pwm_` in hex, must be equal to litexcnc_pwm.h
-    driver_files: ClassVar[List[str]] = [
-        os.path.dirname(__file__) + '/../../driver/modules/litexcnc_pwm.c',
-        os.path.dirname(__file__) + '/../../driver/modules/litexcnc_pwm.h'
-    ]
-    instances: List[PWM_Instance] = Field(
-        [],
-        item_type=PWM_Instance,
-        unique_items=True
-    )
-
-    def create_from_config(self, soc, watchdog):
-        PwmPdmModule.create_from_config(soc, watchdog, self)
-    
-    def add_mmio_config_registers(self, mmio):
-        # The PWM does not require any config, so this function is
-        # not implemented.
-        return
-
-    def add_mmio_write_registers(self, mmio):
-        PwmPdmModule.add_mmio_write_registers(mmio, self)
-
-    def add_mmio_read_registers(self, mmio):
-        # The PWM does not require any read, so this function is
-        # not implemented.
-        return
-
-    @property
-    def config_size(self):
-        return 4
-
-    def store_config(self, mmio):
-        mmio.pwm_config_data =  CSRStatus(
-            size=self.config_size*8,
-            reset=len(self.instances),
-            description=f"The config of the GPIO module."
-        )
